test(gateway): split external agent registry test
